FIX: User ,Aritst, LikeArtist
FIX: User ,Aritst, LikeArtist

close #110

+++ src/apis/artists/artists.resolver.ts
@@ -1,10 +1,7 @@
-import {
-  ConflictException,
-  NotFoundException,
-  UseGuards,
-} from '@nestjs/common';
+import { ConflictException, UseGuards } from '@nestjs/common';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { GqlAuthAccessGuard } from 'src/commons/gql-auth.guard';
+import { RoleService } from 'src/commons/role/role.service';
 import { Roles } from 'src/commons/role/roles.decorator';
 import { RolesGuard } from 'src/commons/role/roles.guard';
 import { RoleType } from 'src/commons/role/type/role-type';
@@ -16,7 +13,10 @@ import { Artist } from './entity/artist.entity';
 
 @Resolver()
 export class ArtistsResolver {
-  constructor(private readonly artistsService: ArtistsService) {}
+  constructor(
+    private readonly artistsService: ArtistsService,
+    private readonly roleService: RoleService,
+  ) {}
 
   @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Artist)
@@ -24,48 +24,58 @@ export class ArtistsResolver {
     @CurrentUser() currentUser: any,
     @Args('createArtistInput') createArtistInput: CreateArtistInput,
   ) {
-    const artist = await this.artistsService.findOneWithActiveName({
+    const artist_ = await this.artistsService.findOneWithActiveName({
       active_name: createArtistInput.active_name,
     });
-    if (artist) {
+    if (artist_) {
       throw new ConflictException('already exist artist');
     }
-
-    return await this.artistsService.create({
+    const artist = await this.artistsService.create({
       ...createArtistInput,
-      userId: currentUser.id,
     });
+    await this.roleService.update({
+      userId: currentUser.id,
+      artistId: artist.id,
+      authority: RoleType.ARTIST,
+    });
+    return artist;
   }
   @Roles(RoleType.ARTIST)
   @UseGuards(RolesGuard)
   @UseGuards(GqlAuthAccessGuard)
   @Query(() => Artist)
-  async fetchArtist(@Args('artistId') artistId: string) {
-    const artist = await this.artistsService.findOne({ artistId });
-    if (!artist) {
-      throw new NotFoundException('not found artist');
-    }
-    return artist;
+  async fetchArtist(@CurrentUser() currentUser) {
+    const role = await this.roleService.findOneWithUserId({
+      userId: currentUser.id,
+    });
+    return await this.artistsService.findOne({ artistId: role.artistId });
   }
 
+  @Roles(RoleType.ARTIST)
+  @UseGuards(RolesGuard)
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
   async updateArtist(
-    @Args('artistId') artistId: string,
+    @CurrentUser() currentUser,
     @Args('updateArtistInput') updateArtistInput: UpdateArtistInput,
   ) {
-    const artist = await this.artistsService.findOne({ artistId });
-    if (!artist) {
-      throw new NotFoundException('not found artist');
-    }
-    return await this.artistsService.update({ artistId, ...updateArtistInput });
+    const role = await this.roleService.findOneWithUserId({
+      userId: currentUser.id,
+    });
+    return await this.artistsService.update({
+      artistId: role.artistId,
+      ...updateArtistInput,
+    });
   }
 
+  @Roles(RoleType.ARTIST)
+  @UseGuards(RolesGuard)
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
-  async deleteArtist(@Args('artistId') artistId: string) {
-    const artist = await this.artistsService.findOne({ artistId });
-    if (!artist) {
-      throw new NotFoundException('not found artist');
-    }
-    return await this.artistsService.delete({ artistId });
+  async deleteArtist(@CurrentUser() currentUser) {
+    const role = await this.roleService.findOneWithUserId({
+      userId: currentUser.id,
+    });
+    return await this.artistsService.delete({ artistId: role.artistId });
   }
 }
